fix(backend): corrige as 71 medias da auditoria SWE (workflow particionado)
Aplicado por 10 agentes em arquivos disjuntos; verificado centralizado (check + 213 testes + makemigrations --check limpo). Temas:
- N+1 eliminados: verificar_pacotes_expirando, verificar_finalizacao, dashboards (agregacao unica em vez de count() em loop)
- transacoes/locks: registrar_falta + verificar_finalizacao com atomic+select_for_update; _transicionar com savepoint p/ auditoria
- I/O externo (push/gcal) movido p/ transaction.on_commit (signals); envios em loop com try/except por item (nao reprocessa batch no retry de Celery)
- timezone: BloqueioAgenda DTSTART em UTC, fila_espera em hora local (off-by-one)
- validacao: MinValueValidator/CheckConstraint em valor_pago + nota (migração 0039)
- except amplo -> especifico/logado; PII mascarada em logs; dead code/imports removidos
- settings: ALLOWED_HOSTS/CSRF strip+filter, DB conn_health_checks+ssl_require, warning REDIS em prod
- termo_obrigado: UX de link expirado (pareado com admin_management)
- bug real exposto por teste fortalecido: booking de data PASSADA era aceito -> +guard 'data futura' em booking_public
- skips seguros reportados pelos agentes (refactors grandes/decisoes de negocio: __str__ PII com teste, Carteira metodos, retry policy de email)

+++ aranha_estetica/services/agendamento.py
@@ -1,6 +1,7 @@
 """Servico de agendamento — orquestra criacao, cancelamento, reagendamento."""
 import logging
 from datetime import timedelta
+from decimal import Decimal
 from typing import Optional
 
 from django.db import IntegrityError, transaction
@@ -25,8 +26,8 @@ class AgendamentoService:
         profissional: Profissional,
         procedimento: Procedimento,
         data_hora_inicio,
-        valor_cobrado: Optional[float] = None,
-        valor_original: Optional[float] = None,
+        valor_cobrado: Optional[Decimal] = None,
+        valor_original: Optional[Decimal] = None,
         descricao_preco: str = "",
         status: str = 'PENDENTE',
     ) -> Atendimento:
